ferm-new/cloud-functions/functions/load_area/requirements.txt
google-cloud-bigquery==3.0.1

firebase-admin==5.2.0

Fiona==1.8.21
Werkzeug==2.1.1
